\name{sgcca-package}
\alias{sgcca-package}
\alias{sgcca}
\docType{package}
\title{
\packageTitle{sgcca}
}
\description{
\packageDescription{sgcca}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{sgcca}
\packageIndices{sgcca}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{sgcca}

Maintainer: \packageMaintainer{sgcca}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{}
